Skeleton refactor: Put scales creation in the Skeleton class
