+++ packages/core/src/components/graphs/skeleton-grid.ts
@@ -3,13 +3,10 @@ import { Skeleton } from "./skeleton";
 import { DOMUtils } from "../../services";
 
 // D3 Imports
-import { scaleLinear } from "d3-scale";
 import { axisLeft, axisBottom } from "d3-axis";
 
 export class SkeletonGrid extends Skeleton {
 	backdrop: any;
-	xScale: any;
-	yScale: any;
 
 	render() {
 		const areDataEmpty = this.model.isDataEmpty();
@@ -24,6 +21,7 @@ export class SkeletonGrid extends Skeleton {
 	}
 
 	renderSkeleton(animate = true) {
+		this.setScales();
 		this.drawBackdrop();
 		this.drawXGrid(animate);
 		this.drawYGrid(animate);
@@ -39,12 +37,8 @@ export class SkeletonGrid extends Skeleton {
 			.attr("width", "100%")
 			.attr("height", "100%");
 
-		const xRange = this.services.cartesianScales.getMainXScale().range();
-		const yRange = this.services.cartesianScales.getMainYScale().range();
-		this.xScale = scaleLinear().domain([0, 1]).range(xRange);
-		this.yScale = scaleLinear().domain([0, 1]).range(yRange);
-		const [xScaleStart, xScaleEnd] = xRange;
-		const [yScaleEnd, yScaleStart] = yRange;
+		const [xScaleStart, xScaleEnd] = this.xScale.range();
+		const [yScaleEnd, yScaleStart] = this.yScale.range();
 
 		this.backdrop
 			.merge(backdropRect)
