+++ packages/core/src/components/graphs/skeleton.ts
@@ -1,8 +1,20 @@
 // Internal Imports
 import { Component } from "../component";
 
+// D3 Imports
+import { scaleLinear } from "d3-scale";
+
 export class Skeleton extends Component {
 	type = "skeleton";
+	xScale: any;
+	yScale: any;
+
+	setScales() {
+		const xRange = this.services.cartesianScales.getMainXScale().range();
+		const yRange = this.services.cartesianScales.getMainYScale().range();
+		this.xScale = scaleLinear().domain([0, 1]).range(xRange);
+		this.yScale = scaleLinear().domain([0, 1]).range(yRange);
+	}
 
 	setStyle() {
 		const svg = this.parent;
